requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
filelock==3.12.3
fsspec==2023.9.0
ftfy==6.1.1
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.3.0
pytesseract==0.3.10
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.32.1
typing_extensions==4.7.1
urllib3==2.0.4
wcwidth==0.2.6
fastapi
uvicorn
aiofiles
python-multipart
aiocache
aiocache[redis]
aioredis==1.3.1
python-dotenv